chore: add groq support
